requirements.txt
asgiref==3.7.2
astroid==3.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2  
colorama==0.4.6
dill==0.3.8
Django==5.0.3
django-environ==0.11.2     
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==24.2.0
google-api-core==2.17.1    
google-auth==2.28.2        
google-geo-type==0.3.7     
google-maps-routing==0.6.7 
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pep8==1.7.1
platformdirs==4.2.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1